- updated VisualStudio version to VS2026 in all exporters

+++ hi_dsp_library/node_api/helpers/ParameterData.cpp
@@ -538,7 +538,7 @@ namespace parameter
 
 		mos.flush();
 
-		parameterNames = new RefCountedHeapBuffer(mos.getDataSize());
+		parameterNames = new RefCountedHeapBuffer((int)mos.getDataSize());
 
 		memcpy(parameterNames->getData(), mos.getData(), mos.getDataSize());
